Coditos

+++ eliminar_usuario.php
@@ -1,8 +1,17 @@
 <?php
-require_once 'conexion.php';
+session_start(); // Inicia la sesión al principio del archivo
 
+// Verifica si no hay un usuario en la sesión o si no es admin
+if(!isset($_SESSION['usuario']) || ($_SESSION['esAdmin'] != 1)) {
+    header("Location: login.html");
+    exit();
+}
+
+// Resto de tu código aquí
 if (isset($_GET['id'])) {
     $id = $_GET['id'];
+    
+    require_once 'conexion.php';
 
     $resultado = eliminarUsuario($id);
